Relocate files and fix test cases
