+++ lib/checkout.rb
