+++ lib/discount_rule.rb
