+++ lib/item.rb
