+++ lib/product.rb
