+++ lib/two_for_one_rule.rb
